requirements.txt
# ABOUTME: Lists Python dependencies required for the page generator and automation tools.
# ABOUTME: Used for environment setup and CI/CD dependency installation.

annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
axe-playwright-python==0.1.7
axe-selenium-python==2.1.6
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
coverage==7.13.4
cryptography==46.0.7
dataclasses-json==0.5.7
distro==1.9.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-genai==1.69.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
greenlet==3.3.2
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
libcst==1.8.6
marshmallow==3.26.2
marshmallow-enum==1.5.1
mypy_extensions==1.1.0
outcome==1.3.0.post0
packaging==26.0
pillow==12.2.0
playwright==1.58.0
pluggy==1.6.0
proto-plus==1.27.1
protobuf==5.29.6
psutil==7.2.2
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.1
Pygments==2.20.0
pyparsing==3.3.2
pyre-check==0.9.25
pyre-extensions==0.0.32
PySocks==1.7.1
pytest==9.0.3
pytest-base-url==2.1.0
pytest-cov==7.0.0
pytest-mock==3.15.1
pytest-playwright==0.7.2
python-dotenv==1.2.2
python-slugify==8.0.4
PyYAML==6.0.3
requests==2.33.1
selenium==4.41.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
tabulate==0.9.0
tenacity==9.1.4
TestSlide==2.7.1
text-unidecode==1.3
tomli==2.4.0
tomli_w==1.2.0
tqdm==4.67.3
trio==0.33.0
trio-websocket==0.12.2
typeguard==2.13.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
webdriver-manager==4.0.2
websocket-client==1.9.0
websockets==16.0
wsproto==1.3.2
